fixed errors

+++ src/main/java/com/suman/airBnbApp/service/HotelServiceImpl.java
@@ -5,12 +5,15 @@ import com.suman.airBnbApp.dto.HotelInfoDto;
 import com.suman.airBnbApp.dto.RoomDto;
 import com.suman.airBnbApp.entity.Hotel;
 import com.suman.airBnbApp.entity.Room;
+import com.suman.airBnbApp.entity.User;
 import com.suman.airBnbApp.exception.ResourceNotFoundException;
+import com.suman.airBnbApp.exception.UnAuthorisedException;
 import com.suman.airBnbApp.repository.HotelRepository;
 import com.suman.airBnbApp.repository.RoomRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -21,27 +24,22 @@ import java.util.List;
 @RequiredArgsConstructor
 public class HotelServiceImpl implements HotelService{
 
-    private final RoomRepository roomRepository;
     private final HotelRepository hotelRepository;
     private final ModelMapper modelMapper;
     private final InventoryService inventoryService;
-
-    @Override
-    public List<HotelDto> getAllHotel() {
-        List<Hotel> hotels = hotelRepository.findAll();
-        List<HotelDto> hotelDtos = hotels.stream()
-                .map((element) -> modelMapper.map(element, HotelDto.class))
-                .toList();
-        return hotelDtos;
-    }
+    private final RoomRepository roomRepository;
 
     @Override
     public HotelDto createNewHotel(HotelDto hotelDto) {
         log.info("Creating a new hotel with name: {}", hotelDto.getName());
         Hotel hotel = modelMapper.map(hotelDto, Hotel.class);
         hotel.setActive(false);
+
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        hotel.setOwner(user);
+
         hotel = hotelRepository.save(hotel);
-        log.info("Created a new hotel with ID: {}", hotel.getId());
+        log.info("Created a new hotel with ID: {}", hotelDto.getId());
         return modelMapper.map(hotel, HotelDto.class);
     }
 
@@ -50,7 +48,12 @@ public class HotelServiceImpl implements HotelService{
         log.info("Getting the hotel with ID: {}", id);
         Hotel hotel = hotelRepository
                 .findById(id)
-                .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with ID: " + id));
+                .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with ID: "+id));
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        if(!user.equals(hotel.getOwner())) {
+            throw new UnAuthorisedException("This user does not own this hotel with id: "+id);
+        }
 
         return modelMapper.map(hotel, HotelDto.class);
     }
@@ -60,23 +63,33 @@ public class HotelServiceImpl implements HotelService{
         log.info("Updating the hotel with ID: {}", id);
         Hotel hotel = hotelRepository
                 .findById(id)
-                .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with ID: " + id));
+                .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with ID: "+id));
+
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner())) {
+            throw new UnAuthorisedException("This user does not own this hotel with id: "+id);
+        }
+
         modelMapper.map(hotelDto, hotel);
         hotel.setId(id);
         hotel = hotelRepository.save(hotel);
         return modelMapper.map(hotel, HotelDto.class);
-
     }
 
     @Override
     @Transactional
     public void deleteHotelById(Long id) {
-        log.info("Delete hotel usinh ID: {}", id);
         Hotel hotel = hotelRepository
                 .findById(id)
-                .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with ID: " + id));
+                .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with ID: "+id));
 
-        for (Room room : hotel.getRooms()){
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner())) {
+            throw new UnAuthorisedException("This user does not own this hotel with id: "+id);
+        }
+
+
+        for(Room room: hotel.getRooms()) {
             inventoryService.deleteAllInventories(room);
             roomRepository.deleteById(room.getId());
         }
@@ -85,20 +98,27 @@ public class HotelServiceImpl implements HotelService{
 
     @Override
     @Transactional
-    public void activeHotel(Long hotelId) {
+    public void activateHotel(Long hotelId) {
         log.info("Activating the hotel with ID: {}", hotelId);
         Hotel hotel = hotelRepository
                 .findById(hotelId)
-                .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with ID: " + hotelId));
-        hotel.setActive(true);
+                .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with ID: "+hotelId));
 
-//        Assuming do it once
-        for (Room room : hotel.getRooms()){
-            inventoryService.initializeRoomForAYear(room);
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        if(!user.equals(hotel.getOwner())) {
+            throw new UnAuthorisedException("This user does not own this hotel with id: "+hotelId);
         }
 
+        hotel.setActive(true);
+
+        // assuming only do it once
+        for(Room room: hotel.getRooms()) {
+            inventoryService.initializeRoomForAYear(room);
+        }
     }
 
+//    public method
     @Override
     public HotelInfoDto getHotelInfoById(Long hotelId) {
         Hotel hotel = hotelRepository
